DABOM-377 refactor: family_member 커서 reader 공통화
