+++ src/main/java/com/template/worker/jobs/common/reader/ActiveFamilyMemberCursorReader.java
@@ -1,0 +1,81 @@
+package com.template.worker.jobs.common.reader;
+
+import java.util.Collections;
+import java.util.Iterator;
+import java.util.List;
+import java.util.function.BiFunction;
+import java.util.function.IntSupplier;
+
+import org.springframework.batch.core.StepExecution;
+import org.springframework.batch.core.StepExecutionListener;
+import org.springframework.batch.item.ItemReader;
+import org.springframework.jdbc.core.JdbcTemplate;
+
+public class ActiveFamilyMemberCursorReader<T> implements ItemReader<T>, StepExecutionListener {
+
+    private static final String READ_ACTIVE_FAMILY_MEMBER_SQL =
+            """
+            SELECT id, family_id, customer_id
+            FROM family_member
+            WHERE deleted_at IS NULL
+              AND id > ?
+            ORDER BY id ASC
+            LIMIT ?
+            """;
+
+    private final JdbcTemplate jdbcTemplate;
+    private final IntSupplier dbFetchSizeSupplier;
+    private final BiFunction<Long, Long, T> targetFactory;
+
+    private Iterator<T> currentBatch = Collections.emptyIterator();
+    private long lastFamilyMemberId = 0L;
+
+    public ActiveFamilyMemberCursorReader(
+            JdbcTemplate jdbcTemplate,
+            IntSupplier dbFetchSizeSupplier,
+            BiFunction<Long, Long, T> targetFactory) {
+        this.jdbcTemplate = jdbcTemplate;
+        this.dbFetchSizeSupplier = dbFetchSizeSupplier;
+        this.targetFactory = targetFactory;
+    }
+
+    @Override
+    public void beforeStep(StepExecution stepExecution) {
+        // Step 재실행 시 상태 누수를 막기 위해 커서를 초기화함
+        currentBatch = Collections.emptyIterator();
+        lastFamilyMemberId = 0L;
+    }
+
+    @Override
+    public T read() {
+        while (!currentBatch.hasNext()) {
+            // PK 커서 기반으로 활성 구성원 목록을 배치 단위 조회함
+            List<FamilyMemberRow> familyMembers =
+                    jdbcTemplate.query(
+                            READ_ACTIVE_FAMILY_MEMBER_SQL,
+                            (resultSet, rowNum) ->
+                                    new FamilyMemberRow(
+                                            resultSet.getLong("id"),
+                                            resultSet.getLong("family_id"),
+                                            resultSet.getLong("customer_id")),
+                            lastFamilyMemberId,
+                            dbFetchSizeSupplier.getAsInt());
+
+            // 더 이상 데이터가 없으면 null을 반환해 Chunk 처리를 종료함
+            if (familyMembers.isEmpty()) {
+                return null;
+            }
+
+            lastFamilyMemberId = familyMembers.get(familyMembers.size() - 1).id();
+            currentBatch =
+                    familyMembers.stream()
+                            .map(row -> targetFactory.apply(row.familyId(), row.customerId()))
+                            .toList()
+                            .iterator();
+        }
+
+        return currentBatch.next();
+    }
+
+    private record FamilyMemberRow(Long id, Long familyId, Long customerId) {}
+}
